pyproject.toml
[tool.poetry]
name = "fastapi-metrics"
version = "0.1.0"
description = ""
authors = ["adryell <adryellpaulo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opentelemetry-api = "^1.29.0"
opentelemetry-sdk = "^1.29.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
opentelemetry-distro = "^0.50b0"
pip = "^24.3.1"
opentelemetry-exporter-otlp = "^1.29.0"
opentelemetry-exporter-prometheus = "^0.50b0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
